+++ resolwe_bio/tests/processes/test_clustering.py
@@ -6,9 +6,9 @@ class ClusteringProcessorTestCase(BioProcessTestCase):
 
     def test_hc_clustering(self):
         """Cannot use assertJSON - JSON output contains ETC object IDs."""
-        expression_1 = self.prepare_expression(f_rc='exp_1_rc.tab.gz', f_exp='exp_1_tpm.tab.gz', f_type="TPM")
-        expression_2 = self.prepare_expression(f_rc='exp_2_rc.tab.gz', f_exp='exp_2_tpm.tab.gz', f_type="TPM")
-        expression_3 = self.prepare_expression(f_rc='exp_2_rc.tab.gz', f_exp='exp_2_tpm.tab.gz', f_type="TPM")
+        expression_1 = self.prepare_expression(f_rc='exp_1_rc.tab.gz', f_exp='exp_1_tpm.tab.gz', f_type='TPM')
+        expression_2 = self.prepare_expression(f_rc='exp_2_rc.tab.gz', f_exp='exp_2_tpm.tab.gz', f_type='TPM')
+        expression_3 = self.prepare_expression(f_rc='exp_2_rc.tab.gz', f_exp='exp_2_tpm.tab.gz', f_type='TPM')
 
         inputs = {'expressions': [expression_1.pk, expression_2.pk, expression_3.pk]}
         etc = self.run_process('etc-bcm', inputs)
@@ -19,23 +19,18 @@ class ClusteringProcessorTestCase(BioProcessTestCase):
         self.run_process('clustering-hierarchical-genes-etc', inputs)
 
     def test_hc_clustering_samples_ucsc(self):
-        inputs = {'exp': 'clustering_expressions_1.tab.gz',
-                  'exp_type': 'Log2',
-                  'exp_name': 'Expression',
-                  'source': 'UCSC'}
-        expression_1 = self.run_process('upload-expression', inputs)
-
-        inputs = {'exp': 'clustering_expressions_2.tab.gz',
-                  'exp_type': 'Log2',
-                  'exp_name': 'Expression',
-                  'source': 'UCSC'}
-        expression_2 = self.run_process('upload-expression', inputs)
-
-        inputs = {'exp': 'clustering_expressions_3.tab.gz',
-                  'exp_type': 'Log2',
-                  'exp_name': 'Expression',
-                  'source': 'UCSC'}
-        expression_3 = self.run_process('upload-expression', inputs)
+        expression_1 = self.prepare_expression(f_exp='clustering_expressions_1.tab.gz',
+                                               f_type='Log2',
+                                               name='Expression',
+                                               source='UCSC')
+        expression_2 = self.prepare_expression(f_exp='clustering_expressions_2.tab.gz',
+                                               f_type='Log2',
+                                               name='Expression',
+                                               source='UCSC')
+        expression_3 = self.prepare_expression(f_exp='clustering_expressions_3.tab.gz',
+                                               f_type='Log2',
+                                               name='Expression',
+                                               source='UCSC')
 
         inputs = {'exps': [expression_1.pk, expression_2.pk, expression_3.pk],
                   'genes': ['A1BG', 'E2f4', 'A2ML1', 'A2MP1', 'A3GALT2', 'A4GALT',
@@ -49,23 +44,18 @@ class ClusteringProcessorTestCase(BioProcessTestCase):
         self.assertTrue('order' in test_json)
 
     def test_hc_clustering_genes_ucsc(self):
-        inputs = {'exp': 'clustering_expressions_1.tab.gz',
-                  'exp_type': 'Log2',
-                  'exp_name': 'Expression',
-                  'source': 'UCSC'}
-        expression_1 = self.run_process('upload-expression', inputs)
-
-        inputs = {'exp': 'clustering_expressions_2.tab.gz',
-                  'exp_type': 'Log2',
-                  'exp_name': 'Expression',
-                  'source': 'UCSC'}
-        expression_2 = self.run_process('upload-expression', inputs)
-
-        inputs = {'exp': 'clustering_expressions_3.tab.gz',
-                  'exp_type': 'Log2',
-                  'exp_name': 'Expression',
-                  'source': 'UCSC'}
-        expression_3 = self.run_process('upload-expression', inputs)
+        expression_1 = self.prepare_expression(f_exp='clustering_expressions_1.tab.gz',
+                                               f_type='Log2',
+                                               name='Expression',
+                                               source='UCSC')
+        expression_2 = self.prepare_expression(f_exp='clustering_expressions_2.tab.gz',
+                                               f_type='Log2',
+                                               name='Expression',
+                                               source='UCSC')
+        expression_3 = self.prepare_expression(f_exp='clustering_expressions_3.tab.gz',
+                                               f_type='Log2',
+                                               name='Expression',
+                                               source='UCSC')
 
         inputs = {'exps': [expression_1.pk, expression_2.pk, expression_3.pk],
                   'genes': ['A1BG', 'E2f4', 'A2ML1', 'A2MP1', 'A3GALT2', 'A4GALT',
@@ -79,23 +69,18 @@ class ClusteringProcessorTestCase(BioProcessTestCase):
         self.assertTrue('order' in test_json)
 
     def test_hc_clustering_genes_ncbi(self):
-        inputs = {'exp': 'clustering_NCBI.gz',
-                  'exp_type': 'rc',
-                  'exp_name': 'Expression',
-                  'source': 'NCBI'}
-        expression_1 = self.run_process('upload-expression', inputs)
-
-        inputs = {'exp': 'clustering_NCBI_1.gz',
-                  'exp_type': 'rc',
-                  'exp_name': 'Expression',
-                  'source': 'NCBI'}
-        expression_2 = self.run_process('upload-expression', inputs)
-
-        inputs = {'exp': 'clustering_NCBI_2.gz',
-                  'exp_type': 'rc',
-                  'exp_name': 'Expression',
-                  'source': 'NCBI'}
-        expression_3 = self.run_process('upload-expression', inputs)
+        expression_1 = self.prepare_expression(f_exp='clustering_NCBI.gz',
+                                               f_type='rc',
+                                               name='Expression',
+                                               source='NCBI')
+        expression_2 = self.prepare_expression(f_exp='clustering_NCBI_1.gz',
+                                               f_type='rc',
+                                               name='Expression',
+                                               source='NCBI')
+        expression_3 = self.prepare_expression(f_exp='clustering_NCBI_2.gz',
+                                               f_type='rc',
+                                               name='Expression',
+                                               source='NCBI')
 
         inputs = {'exps': [expression_1.pk, expression_2.pk, expression_3.pk],
                   'genes': ['1', '503538', '56934', '29974', '2', '144571', '3'],
@@ -107,23 +92,18 @@ class ClusteringProcessorTestCase(BioProcessTestCase):
         self.assertTrue('order' in test_json)
 
     def test_hc_clustering_samples_ncbi(self):
-        inputs = {'exp': 'clustering_NCBI.gz',
-                  'exp_type': 'rc',
-                  'exp_name': 'Expression',
-                  'source': 'NCBI'}
-        expression_1 = self.run_process('upload-expression', inputs)
-
-        inputs = {'exp': 'clustering_NCBI_1.gz',
-                  'exp_type': 'rc',
-                  'exp_name': 'Expression',
-                  'source': 'NCBI'}
-        expression_2 = self.run_process('upload-expression', inputs)
-
-        inputs = {'exp': 'clustering_NCBI_2.gz',
-                  'exp_type': 'rc',
-                  'exp_name': 'Expression',
-                  'source': 'NCBI'}
-        expression_3 = self.run_process('upload-expression', inputs)
+        expression_1 = self.prepare_expression(f_exp='clustering_NCBI.gz',
+                                               f_type='rc',
+                                               name='Expression',
+                                               source='NCBI')
+        expression_2 = self.prepare_expression(f_exp='clustering_NCBI_1.gz',
+                                               f_type='rc',
+                                               name='Expression',
+                                               source='NCBI')
+        expression_3 = self.prepare_expression(f_exp='clustering_NCBI_2.gz',
+                                               f_type='rc',
+                                               name='Expression',
+                                               source='NCBI')
 
         inputs = {'exps': [expression_1.pk, expression_2.pk, expression_3.pk],
                   'genes': ['1', '503538', '56934', '29974', '2', '144571', '3'],